Recursive mutex used. InvalidateArea() added.
Rtos and Rtos Wrapper should support recursive mutexes. It used in display driver and allow nested functions to call LockVisObject() multiple times.

Now display updates rectangular area that contains object that changed since last display update. It improve performance in some cases.

+++ Libraries/Tcs34725.cpp
@@ -280,7 +280,7 @@ Result Tcs34725::GetLux(uint16_t& lux)
     float cpl = (((256 - integration_time) * 2.4f) * GetGainValue()) / (1.0f * 310.0f);
 
     // Finally, calculate the lux
-    lux = (0.136f * (float)r + 1.000f * (float)g - 0.444f * (float)b) / cpl;
+    lux = (uint16_t)((0.136f * (float)r + 1.000f * (float)g - 0.444f * (float)b) / cpl);
 
     result = Result::RESULT_OK;
   }
